+++ tgpt/__init__.py
@@ -1,7 +1,7 @@
 from .tgpt import TGPT
 from .imager import Imager
 
-__version__ = "0.0.6"
+__version__ = "0.0.7"
 __author__ = "Smartwa"
 __repo__ = "https://github.com/Simatwa/python-tgpt"